src/pyproject.toml
[project]
name = "tbank-converter"
version = "0.1.0"
description = "CLI утилита для конвертации CSV выписок Т-Банка в XLSX формат"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.scripts]
tbank-convert = "tbank_converter.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
